+++ src/main/java/mdteam/ait/registry/CreakRegistry.java
@@ -29,6 +29,11 @@ public class CreakRegistry {
     public static CreakSound ONE;
     public static CreakSound TWO;
     public static CreakSound THREE;
+    // Concerned that 4-7 may be too quiet
+    public static CreakSound FOUR;
+    public static CreakSound FIVE;
+    public static CreakSound SIX;
+    public static CreakSound SEVEN;
     public static CreakSound CAVE;
     public static CreakSound WHISPER;
 
@@ -36,6 +41,11 @@ public class CreakRegistry {
         ONE = register(CreakSound.create(AITMod.MOD_ID, "one", AITSounds.CREAK_ONE));
         TWO = register(CreakSound.create(AITMod.MOD_ID, "two", AITSounds.CREAK_TWO));
         THREE = register(CreakSound.create(AITMod.MOD_ID, "three", AITSounds.CREAK_THREE));
+        FOUR = register(CreakSound.create(AITMod.MOD_ID, "four", AITSounds.CREAK_FOUR));
+        FIVE = register(CreakSound.create(AITMod.MOD_ID, "five", AITSounds.CREAK_FIVE));
+        SIX = register(CreakSound.create(AITMod.MOD_ID, "six", AITSounds.CREAK_SIX));
+        SEVEN = register(CreakSound.create(AITMod.MOD_ID, "seven", AITSounds.CREAK_SEVEN));
+
         CAVE = register(CreakSound.create(AITMod.MOD_ID, "cave", SoundEvents.AMBIENT_CAVE.value()));
         WHISPER = register(CreakSound.create(AITMod.MOD_ID, "whisper", AITSounds.WHISPER));
     }