+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated monitor screen
 - Added eighth doctor hum
 - Made a different particle show when refueling in a RiftChunk
+- Added four new creak sounds
 - Failing to takeoff/land no longer sends a message to the pilot
 - Created ArtronHolder interface
 - Renamed "ExteriorSchema" to "ExteriorCategory"